Moving Authentication package out of internal package
- This is required to run OSGi tests in siddhi-io-wso2event
