+++ components/data-bridge/org.wso2.carbon.databridge.core/src/main/java/org/wso2/carbon/databridge/core/authentication/CarbonAuthenticationHandler.java
@@ -16,7 +16,7 @@
 * under the License.
 */
 
-package org.wso2.carbon.databridge.core.internal.authentication;
+package org.wso2.carbon.databridge.core.authentication;
 
 import org.wso2.carbon.databridge.core.utils.AgentSession;
